29-Nov-2019

+++ Basics/Calculations.cs
@@ -4,8 +4,8 @@ class Calculations
 {
 	static void Main()
 	{
-		int num1 = 34;
-		int num2 = 53;
+		double num1 = 34.0;
+		double num2 = 53.0;
 		
 		Console.WriteLine(num1 + " + " + num2 + " = " + (num1 + num2));
 		Console.WriteLine(num1 + " - " + num2 + " = " + (num1 - num2));
